requirements.txt
tensorboard==1.9.0
tensorflow-gpu==1.9.0
tqdm==4.24.0
numpy==1.15.0
pandas==0.23.4
sklearn
batchup==0.2.0
torchvision==0.2.1